refactor(commands): move orchestration to core services for thin adapters (#5380)
Resolve a coherent subset of the thin_command_adapter violations flagged by
the #4384 detector by moving orchestration out of command adapters into core
services, leaving the command modules to parse args, delegate, and adapt
output. Behavior is unchanged — code is moved, not altered.

- trace/experiment.rs: command/process execution and artifact collection moved
  to new core::trace_experiment service; command now builds the plan and
  delegates run/collect.
- tunnel.rs: preview-consumer execution (mkdir, process spawn, artifact
  persistence, config parsing) moved to new core::preview_consumer service;
  command delegates and adapts the result.
- bench/tests.rs: root-caused false positive — this file is a pure
  `#[cfg(test)] mod tests` module but was scanned as production because
  is_test_path did not recognize bare `tests.rs`/`test.rs` filenames. Extended
  is_test_path so these conventional test modules are skipped by all detectors.

Cleared the three resolved findings from the audit baseline. Remaining #5380
modules (bench/observation.rs, observe.rs, review/mod.rs,
agent_task/controller.rs) are larger orchestration moves left as follow-up.
