+++ homeboy.json
@@ -780,7 +780,6 @@
         "src/commands/agent_task/tool.rs::ThinCommandAdapterViolation::thin_command_adapter::Command module accumulates orchestration that belongs in a core service: runner execution orchestration",
         "src/commands/bench/observation.rs::ThinCommandAdapterViolation::thin_command_adapter::Command module accumulates orchestration that belongs in a core service: direct filesystem mutation, run artifact persistence",
         "src/commands/bench/settings_matrix.rs::ThinCommandAdapterViolation::thin_command_adapter::Command module accumulates orchestration that belongs in a core service: run artifact persistence",
-        "src/commands/bench/tests.rs::ThinCommandAdapterViolation::thin_command_adapter::Command module accumulates orchestration that belongs in a core service: direct filesystem mutation",
         "src/commands/doctor/resources.rs::ThinCommandAdapterViolation::thin_command_adapter::Command module accumulates orchestration that belongs in a core service: direct process execution",
         "src/commands/json_output/mod.rs::ThinCommandAdapterViolation::thin_command_adapter::Command module accumulates orchestration that belongs in a core service: runner execution orchestration",
         "src/commands/observe.rs::ThinCommandAdapterViolation::thin_command_adapter::Command module accumulates orchestration that belongs in a core service: direct filesystem mutation, direct process execution, run artifact persistence",
@@ -793,11 +792,9 @@
         "src/commands/trace/compare_bundle.rs::ThinCommandAdapterViolation::thin_command_adapter::Command module accumulates orchestration that belongs in a core service: direct filesystem mutation",
         "src/commands/trace/compare_targets.rs::ThinCommandAdapterViolation::thin_command_adapter::Command module accumulates orchestration that belongs in a core service: direct filesystem mutation, run artifact persistence, runner execution orchestration",
         "src/commands/trace/compare_variant.rs::ThinCommandAdapterViolation::thin_command_adapter::Command module accumulates orchestration that belongs in a core service: direct filesystem mutation",
-        "src/commands/trace/experiment.rs::ThinCommandAdapterViolation::thin_command_adapter::Command module accumulates orchestration that belongs in a core service: direct filesystem mutation, direct process execution",
         "src/commands/trace/matrix.rs::ThinCommandAdapterViolation::thin_command_adapter::Command module accumulates orchestration that belongs in a core service: direct filesystem mutation, runner execution orchestration",
         "src/commands/trace/observation_lifecycle.rs::ThinCommandAdapterViolation::thin_command_adapter::Command module accumulates orchestration that belongs in a core service: run artifact persistence",
         "src/commands/trace/test_fixture.rs::ThinCommandAdapterViolation::thin_command_adapter::Command module accumulates orchestration that belongs in a core service: direct filesystem mutation, direct process execution",
-        "src/commands/tunnel.rs::ThinCommandAdapterViolation::thin_command_adapter::Command module accumulates orchestration that belongs in a core service: direct filesystem mutation, direct process execution",
         "src/commands/utils/observed_workflow.rs::ThinCommandAdapterViolation::thin_command_adapter::Command module accumulates orchestration that belongs in a core service: run artifact persistence",
         "src/commands/utils/output.rs::ThinCommandAdapterViolation::thin_command_adapter::Command module accumulates orchestration that belongs in a core service: direct file writer construction, direct filesystem mutation",
         "structural::src/commands/agent_task_summary.rs::HighItemCount",
