+++ src/commands/tunnel.rs
@@ -1,6 +1,5 @@
 use clap::{Args, Subcommand, ValueEnum};
-use serde::{Deserialize, Serialize};
-use serde_json::Value;
+use serde::Serialize;
 
 use homeboy::core::artifacts::{
     self, ArtifactOriginInspect, ArtifactOriginServeSpec, ArtifactOriginStatus,
@@ -8,6 +7,7 @@ use homeboy::core::artifacts::{
 use homeboy::core::preview_client::{
     self, PreviewClientAuthDiagnostic, PreviewClientReport, PreviewClientStartSpec,
 };
+use homeboy::core::preview_consumer;
 use homeboy::core::preview_ingress::{
     self, PreviewIngressInstallOptions, PreviewIngressInstallPlan, PreviewIngressInstallStatusPlan,
     PreviewIngressRoute, PreviewIngressServeSpec, PreviewIngressStatus,
@@ -20,9 +20,7 @@ use homeboy::core::tunnel::{
 };
 use homeboy::core::{EntityCrudOutput, MergeOutput};
 use std::collections::BTreeMap;
-use std::fs;
 use std::path::PathBuf;
-use std::process::Command;
 
 use super::{CmdResult, DynamicSetArgs};
 
@@ -68,54 +66,8 @@ pub enum PreviewIngressActionOutput {
     Status(PreviewIngressStatus),
 }
 
-#[derive(Debug, Serialize)]
-pub struct PreviewConsumerOutput {
-    pub schema: String,
-    pub consumer_id: String,
-    pub preview_public_url: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub service_id: Option<String>,
-    pub artifacts_dir: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub public_result_url: Option<String>,
-    pub exit_code: i32,
-    pub stdout: String,
-    pub stderr: String,
-    pub artifact_path: String,
-}
-
-#[derive(Debug, Deserialize)]
-struct PreviewConsumerConfig {
-    pub id: String,
-    pub command: PreviewConsumerCommandConfig,
-    #[serde(default)]
-    pub output: PreviewConsumerOutputConfig,
-    #[serde(default)]
-    pub artifact_file: Option<String>,
-}
-
-#[derive(Debug, Deserialize)]
-struct PreviewConsumerCommandConfig {
-    pub program: String,
-    #[serde(default)]
-    pub args: Vec<String>,
-    #[serde(default)]
-    pub cwd: Option<PathBuf>,
-    #[serde(default)]
-    pub env: BTreeMap<String, String>,
-    #[serde(default)]
-    pub artifacts_dir: Option<PathBuf>,
-}
-
-#[derive(Debug, Default, Deserialize)]
-struct PreviewConsumerOutputConfig {
-    #[serde(default)]
-    pub public_result_json_file: Option<PathBuf>,
-    #[serde(default)]
-    pub public_result_json_pointer: Option<String>,
-    #[serde(default)]
-    pub public_result_stdout_prefix: Option<String>,
-}
+/// Preview-consumer run output, owned by the core preview-consumer service.
+pub use homeboy::core::preview_consumer::PreviewConsumerRunResult as PreviewConsumerOutput;
 
 #[derive(Debug, Serialize)]
 #[serde(tag = "action", rename_all = "snake_case")]
@@ -779,92 +731,17 @@ fn run_preview_consumer_config(
     preview_public_url: Option<String>,
     artifacts_dir_override: Option<PathBuf>,
 ) -> CmdResult<TunnelOutput> {
-    let config = read_preview_consumer_config(&config_path)?;
-    let public_url =
-        resolve_preview_consumer_public_url(service_id.as_deref(), preview_public_url.as_deref())?;
-    let artifacts_dir = artifacts_dir_override
-        .or_else(|| config.command.artifacts_dir.clone())
-        .unwrap_or_else(|| {
-            homeboy::core::artifacts::root()
-                .unwrap_or_else(|_| std::env::temp_dir().join("homeboy-artifacts"))
-                .join("preview-consumer")
-                .join(safe_artifact_slug(&config.id))
-        });
-    fs::create_dir_all(&artifacts_dir).map_err(|err| {
-        homeboy::core::Error::internal_io(
-            err.to_string(),
-            Some(format!("create artifacts dir {}", artifacts_dir.display())),
-        )
-    })?;
-
-    let mut command = Command::new(render_preview_consumer_template(
-        &config.command.program,
-        &public_url,
-        &artifacts_dir,
-    ));
-    for arg in &config.command.args {
-        command.arg(render_preview_consumer_template(
-            arg,
-            &public_url,
-            &artifacts_dir,
-        ));
-    }
-    for (key, value) in &config.command.env {
-        command.env(
-            key,
-            render_preview_consumer_template(value, &public_url, &artifacts_dir),
-        );
-    }
-    if let Some(cwd) = &config.command.cwd {
-        command.current_dir(cwd);
-    }
-
-    let output = command.output().map_err(|err| {
-        homeboy::core::Error::internal_io(
-            err.to_string(),
-            Some(format!("run preview consumer {}", config.id)),
-        )
-    })?;
-    let exit_code = output.status.code().unwrap_or(1);
-    let stdout = String::from_utf8_lossy(&output.stdout).to_string();
-    let stderr = String::from_utf8_lossy(&output.stderr).to_string();
-
-    let public_result_url =
-        extract_preview_consumer_public_result_url(&config.output, &artifacts_dir, &stdout);
-    let artifact_file = config
-        .artifact_file
-        .as_deref()
-        .unwrap_or("homeboy-preview-consumer.json");
-    let result = PreviewConsumerOutput {
-        schema: "homeboy/preview-consumer-run/v1".to_string(),
-        consumer_id: config.id.clone(),
-        preview_public_url: public_url,
+    let (result, exit_code) = preview_consumer::run(preview_consumer::PreviewConsumerRunRequest {
+        config_path,
         service_id,
-        artifacts_dir: artifacts_dir.display().to_string(),
-        public_result_url,
-        exit_code,
-        stdout,
-        stderr,
-        artifact_path: artifacts_dir.join(artifact_file).display().to_string(),
-    };
-
-    let artifact_json = serde_json::to_string_pretty(&result).map_err(|err| {
-        homeboy::core::Error::internal_json(
-            err.to_string(),
-            Some("serialize preview consumer run artifact".to_string()),
-        )
-    })?;
-    fs::write(&result.artifact_path, format!("{artifact_json}\n")).map_err(|err| {
-        homeboy::core::Error::internal_io(
-            err.to_string(),
-            Some(format!("write {}", result.artifact_path)),
-        )
+        preview_public_url,
+        artifacts_dir_override,
     })?;
 
     Ok((
         TunnelOutput {
             command: "tunnel.preview_consumer.run".to_string(),
-            id: Some(config.id),
+            id: Some(result.consumer_id.clone()),
             extra: TunnelExtra {
                 preview_consumer: Some(result),
                 ..Default::default()
@@ -873,113 +750,6 @@ fn run_preview_consumer_config(
         },
         exit_code,
     ))
-}
-
-fn read_preview_consumer_config(
-    path: &std::path::Path,
-) -> homeboy::core::Result<PreviewConsumerConfig> {
-    let raw = fs::read_to_string(path).map_err(|err| {
-        homeboy::core::Error::internal_io(
-            err.to_string(),
-            Some(format!("read preview consumer config {}", path.display())),
-        )
-    })?;
-    serde_json::from_str(&raw).map_err(|err| {
-        homeboy::core::Error::validation_invalid_json(
-            err,
-            Some(format!("parse preview consumer config {}", path.display())),
-            Some(raw),
-        )
-    })
-}
-
-fn resolve_preview_consumer_public_url(
-    service_id: Option<&str>,
-    preview_public_url: Option<&str>,
-) -> homeboy::core::Result<String> {
-    if let Some(public_url) = preview_public_url {
-        return Ok(public_url.to_string());
-    }
-    let Some(service_id) = service_id else {
-        return Err(homeboy::core::Error::validation_missing_argument(vec![
-            "--service-id or --preview-public-url".to_string(),
-        ]));
-    };
-    let status = tunnel::status(service_id)?;
-    status
-        .preview_identity
-        .public_url
-        .or_else(|| status.preview.and_then(|preview| preview.preview_identity.public_url))
-        .ok_or_else(|| {
-            homeboy::core::Error::validation_invalid_argument(
-                "service-id",
-                "service status does not contain a public preview URL; start the service with a public tunnel backend first",
-                Some(service_id.to_string()),
-                None,
-            )
-        })
-}
-
-fn render_preview_consumer_template(
-    value: &str,
-    public_url: &str,
-    artifacts_dir: &std::path::Path,
-) -> String {
-    value
-        .replace("${preview_public_url}", public_url)
-        .replace("${artifacts_dir}", &artifacts_dir.to_string_lossy())
-}
-
-fn extract_preview_consumer_public_result_url(
-    config: &PreviewConsumerOutputConfig,
-    artifacts_dir: &std::path::Path,
-    stdout: &str,
-) -> Option<String> {
-    config
-        .public_result_json_file
-        .as_ref()
-        .and_then(|path| {
-            let path = if path.is_absolute() {
-                path.clone()
-            } else {
-                artifacts_dir.join(path)
-            };
-            let raw = fs::read_to_string(path).ok()?;
-            serde_json::from_str::<Value>(&raw).ok()
-        })
-        .and_then(|value| {
-            config
-                .public_result_json_pointer
-                .as_deref()
-                .and_then(|pointer| json_pointer_string(&value, pointer))
-        })
-        .or_else(|| {
-            config
-                .public_result_stdout_prefix
-                .as_deref()
-                .and_then(|prefix| parse_prefixed_line(stdout, prefix))
-        })
-}
-
-fn json_pointer_string(value: &Value, pointer: &str) -> Option<String> {
-    value.pointer(pointer)?.as_str().map(str::to_string)
-}
-
-fn parse_prefixed_line(output: &str, prefix: &str) -> Option<String> {
-    output.lines().find_map(|line| {
-        line.strip_prefix(prefix)
-            .map(str::trim)
-            .filter(|value| !value.is_empty())
-            .map(str::to_string)
-    })
-}
-
-fn safe_artifact_slug(value: &str) -> String {
-    let slug: String = value
-        .chars()
-        .map(|ch| if ch.is_ascii_alphanumeric() { ch } else { '-' })
-        .collect();
-    slug.trim_matches('-').chars().take(96).collect()
 }
 
 fn run_preview_client(command: TunnelPreviewClientCommand) -> CmdResult<TunnelOutput> {
@@ -1574,24 +1344,6 @@ mod tests {
             assert_eq!(output.command, "tunnel.service.expose");
             assert_eq!(output.entity.expect("entity").id, "runner-local-preview");
         });
-    }
-
-    #[test]
-    fn parses_public_result_url_from_configured_stdout_prefix() {
-        let stdout = "Consumer ready\nPublic result URL: https://run.example.test/result\n";
-
-        assert_eq!(
-            parse_prefixed_line(stdout, "Public result URL:").as_deref(),
-            Some("https://run.example.test/result")
-        );
-    }
-
-    #[test]
-    fn safe_artifact_slug_keeps_consumer_id_human_readable() {
-        assert_eq!(
-            safe_artifact_slug("preview consumer: sample"),
-            "preview-consumer--sample"
-        );
     }
 
     #[test]
